+++ src/main/java/com/bizcub/randomItemSpeedrun/client/gui/ScaledItemPIPRenderer.java
@@ -12,7 +12,7 @@ import net.minecraft.client.gui.navigation.ScreenRectangle;
 import net.minecraft.client.gui.render.TextureSetup;
 import net.minecraft.client.gui.render.pip.PictureInPictureRenderer;
 /*? <26.2*/ import net.minecraft.client.renderer.MultiBufferSource;
-import net.minecraft.client.renderer.SubmitNodeStorage;
+/*? >=26.2*/ //import net.minecraft.client.renderer.SubmitNodeStorage;
 import net.minecraft.client.renderer.state.gui.BlitRenderState;
 import net.minecraft.client.renderer.state.gui.GuiRenderState;
 import net.minecraft.client.renderer.RenderPipelines;