+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-* The mod now works on the server as well.
-* Added a sound when an item appears on the screen.
-* Changed the sound for successfully completing a game.
+* Item rendering has been added to the HUD
+* An option to remove duplicates has been added